requirements.txt
ultralytics 
numpy
matplotlib
pandas==1.5.0
Pillow
folium==0.14
geopy